Add files via upload
This is the "latest" version. I fixed many of the "bugs" in it but it only reduces the score. I have tried running it at lower depths so it is less prone to crashing (changing the check for 30 marbles or less into 26 marbles or less) but it does not help too much.